requirements.txt
mysql-connector-python
sentence-transformers
numpy
fastapi 
uvicorn 